backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.7.2
chromadb==0.5.5  # Cập nhật lên phiên bản mới
python-dotenv==1.0.0
numpy==1.26.4
# pip install python-docx
